Remove 'FINAL' keyword because we don't need it (#2)
Users reported issues when using `final` as the name of a CTE, which was caused because sqlparse treats it as a reserved keyword. It appears to me that this is only a keyword in DB2 SQL, which I don't think we care much about, so I've simply removed it from the keyword list.